Correct the man pages install location
In 1c58e3ba99a64948bda6aaf1a21220cbe8741880 on CMake 3.14 or later,
we attempted to specify TYPE MAN for the manual pages.
There does not appear to be a way to specify a subdirectory
when using that notation. Let us revert to the older INSTALL
variant with explicit DESTINATION.

This error was pointed out by László Böszörményi (GCS).